requirements.txt
jupyter-ydoc
nbclient
nbformat
aiofiles
aiohttp
websockets
mcp
pytest
pytest-asyncio
python-dotenv